advanced-software-engineering/Authentication-2/jwt_demo/requirements.txt
Django>=4.2
PyJWT>=2.8.0
